requirements-prod.txt
fastapi==0.110.0
gunicorn==21.2.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
numpy==1.26.4
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
qrcode==7.4.2
pillow==10.2.0
opencv-python-headless==4.9.0.80
email-validator==2.1.1
pandas==2.2.0
openpyxl==3.1.2
python-dotenv==1.0.1
telethon==1.34.0
pydantic-settings==2.2.1
